Cleaned up concatenation of bytes into bigger integers
- Whitespace cleanup
- Made all casts use uint32_t during combining. The functions returning int32_t
  will then cast that uint32_t as part of the return implicitly.

+++ tmc/ic/TMC2209/TMC2209.c
@@ -165,7 +165,10 @@ int32_t readRegisterUART(uint16_t icID, uint8_t address)
     if (data[7] != CRC8(data, 7))
         return 0;
 
-    return ((uint32_t) data[3] << 24) | ((uint32_t) data[4] << 16) | ((uint32_t) data[5] << 8) | data[6];
+    return    ((uint32_t) data[3] << 24)
+            | ((uint32_t) data[4] << 16)
+            | ((uint32_t) data[5] << 8)
+            | ((uint32_t) data[6]);
 }
 
 void writeRegisterUART(uint16_t icID, uint8_t address, int32_t value)
